fix: refactor home page component
